requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.13.1
fsspec==2024.6.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==2.1.2
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
pandas==2.3.3
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
safetensors==0.6.1
sentry-sdk==2.51.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.13.3
tiktoken==0.10.0
tokenizers==0.21.4
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
